refactor: reorganize repository structure and remove unused repository interfaces
