+++ src/main/java/org/kungnection/repository/FriendshipDAO.java
@@ -1,5 +1,6 @@
-package org.kungnection.db;
+package org.kungnection.repository;
 
+import org.kungnection.db.DatabaseUtil;
 import org.kungnection.model.Friendship;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
