--- src/main/java/org/kungnection/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package org.kungnection.repository;
-
-import org.kungnection.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    User findByEmail(String email);
-
-    User findByUsername(String username);
-}
